Support OP-TEE in generic TEE subsystem
Replaces support for for the old OP-TEE driver with support for OP-TEE
in the generic TEE subsystem.

Uses id instead of fd to identify shared memory objects to let
tee-supplicant manage allocation and freeing of its memory shared with
secure world.

This change depends on corresponding abi change in the kernel.

Reviewed-by: Joakim Bech <joakim.bech@linaro.org>
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>