Try loading TAs from another writable path
To allow testing of corrupt TA loading on systems where the TA
directory is read-only, try loading TAs first from an alternate
`TEEC_TEST_LOAD_PATH` before the regular path.  This allows the test
to override the TA with one for testing purposes.

Although useful for testing, this should not be enabled for a
production system, since it facilitates replacing a TA with a corrupt
image.  CFG_TA_TEST_PATH can be turned off in the `config.mk`.

Signed-off-by: David Brown <david.brown@linaro.org>
Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>