[fix] libsks: implement C_Encrypt, C_Decrypt and friends
Client can call cipher update/final without an output buffer reference.

The library shall relay to the SKS TA the input arguments provided by
the client. If the client does not provide an output data size reference
where SKS could check/set to output data length, relay this to the
TA so that it can decide how to handle such an invocation.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>