[fix] libsks: C_Sign/C_Verify and friends for AES flavored MAC
Client can call sing/verify final without an output buffer reference.

The library shall relay to the SKS TA the input arguments provided by
the client. If the client does not provide an output data size reference
where SKS could check/set to output data length, relay this to the
TA so that it can decide how to handle such an invocation.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>