libteec: Makefile: Generate .so files the same way as CMake
Update the Makefile so that the generated libteec.so file and the
corresponding symlinks match those generated by the CMake script.

Specifically, before this commit the Makefile generated the following
files:
  symlink: libteec.so -> libteec.so.1
  symlink: libteec.so.1 -> libteec.so.1.0
  file: libteec.so.1.0 (SONAME libteec.so.1.0)

With this commit the generated files are:
  symlink: libteec.so -> libteec.so.1
  symlink: libteec.so.1 -> libteec.so.1.0.0
  symlink: libteec.so.1.0 -> libteec.so.1.0.0
  file: libteec.so.1.0.0 (SONAME libteec.so.1)

The symlink "libteec.so.1.0" is kept for compatibility with existing
installations. Note that "libteec.so.1.0" is not generated by the CMake
script.

Signed-off-by: Fabian Hofmann <fabian.hofmann@fau.de>
Acked-by: Jerome Forissier <jerome.forissier@linaro.org>