libutee: flag NULL pointer using invalid shm id
Referring to GPTS test suite v2.0.0.1-2016-11-09 id de-14-33

Refer to the TEEC_RegisterSharedMemory function for error conditions
which can be triggered during temporary registration of a memory region.

The CA needs to be able to send a null pointer to the TA. Previously
this was done by allocating a block of shared memory but setting the
size to 0.  This did not allow sending a non-NULL pointer with size 0,
this failing GPTS tests.
This change flags a NULL pointer with an invalid shared memory id rather
than size 0.

Add new context capability "memref_null" in the TEEC_Context
structure to handle the support of the shared memory buffer
with a NULL pointer.

Acked-by: Etienne Carriere <etienne.carriere@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>
Signed-off-by: Michael Whitfield <michael.whitfield@nxp.com>
Signed-off-by: Cedric Neveux <cedric.neveux@nxp.com>