libckteec: helper function to serialize a attribute template
Helper function serialize_ck_attributes() takes an attribute list
and generates a serial representation of it.

The template contains item that can be of a given size or that can
be a list of template, referred to as indirect attributes.

In the process, the Cryptoki ULONG data are serialized as 32bit
data as per PKCS11 TA ABI.

Acked-by: Jerome Forissier <jerome@forissier.org>
Co-developed-by: Etienne Carriere <etienne.carriere@linaro.org>
Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>