libteec: fix TEEC_RegisterSharedMemory() with a fallback option
Adds a fallback option to TEEC_RegisterSharedMemory() where
TEE_IOC_SHM_REGISTER fails, likely due to the caller supplying read-only
memory. Instead of failing and burden the caller to handle this,
implement it here instead. If TEE_IOC_SHM_REGISTER fails, malloc() a
buffer and register that as a shadow buffer instead.

The size of the TEEC_SharedMemory is kept unchanged, but the last field
bool buffer_allocated is replaced with a union of the same size but with
a flags field which will be used instead. This allows keeping track of
two internal flags, SHM_FLAG_BUFFER_ALLOCED and
SHM_FLAG_SHADOW_BUFFER_ALLOCED while avoiding changing the ABI. There
are a few bits left for future use.

Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>
Reviewed-by: Jerome Forissier <jerome@forissier.org>
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>