Add ckteecpp Interface library
Add a number of class definitions to ensure that the cleanup functions
are called after use.

In C++, one of the easiest ways to ensure that cleanup functions are
called once in both normal path and exception path is to allocate an
object with a destructor on the stack. The classes added here are
minimal and the return value handling should be thought further.

Signed-off-by: Eero Aaltonen <eero.aaltonen@vaisala.com>
Acked-by: Etienne Carriere <etienne.carriere@linaro.org>