Makefile: install optee_client_config.mk at INCLUDEDIR
optee_client_config.mk should be installed to a more traditional and
common folder such as INCLUDEDIR instead of being installed directly to
DESTDIR (which can end up as / depending on the packaging system used).

This was noticed when packaging optee-client for OE, as DESTDIR gets set
to the deploy folder, which is later used as root for packaging
optee-client.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>
Acked-by: Jerome Forissier <jerome@forissier.org>