libteec: Move OP-TEE defined fields into an imp struct
GlobalPlatform TEE Client API Specification v1.0 specifies that
the structs TEEC_Context, TEEC_Session, TEEC_SharedMemory,
and TEEC_Operation shall have a user defined struct named imp.
In OP-TEE the struct is not there and instead the user defined
fields are declared directly in the top structs.
This commit introduces the imp struct to better support using
different implementations. The imp fields now represent the
implementation defined parts of the structs that was
previously declared directly in the top struct. All previously
available parameters are preserved in the imp struct.
The updated version of the imp structure makes it easier to
create a binding for Rust.
Adding the missing imp struct to the structs in OP-TEE is an
ABI breakage which requires a version major update of libteec
as explained by Jerome Forissier.

Link: https://github.com/OP-TEE/optee_client/issues/348
Reported-by: Tom Hebb <tommyhebb@gmail.com>
Signed-off-by: Julianus Larson <julianus.larson@linaro.org>
Acked-by: Etienne Carriere <etienne.carriere@foss.st.com>
Acked-by: Jerome Forissier <jerome.forissier@linaro.org>

Fixed spacing

Signed-off-by: Julianus Larson <julianus.larson@linaro.org>