tee-supplicant: Allow for TA load path to be specified at runtime
Add a new `--ta-path` CLI flag for overriding the default load path used
by tee-supplicant. The given path string can be a set of colon (':')
separated paths, each being a full path used when searching for TAs.
When this option is not used, the existing behavior of loading TAs from
a subdirectory "ta-dir" under TEEC_LOAD_PATH is retained.

Signed-off-by: Jared Baur <jaredbaur@fastmail.com>
Reviewed-by: Etienne Carriere <etienne.carriere@foss.st.com>
Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>