tee-supplicant: move a non-working fs_fsync() call
fs_fsync() was called in the do_mkdir function that is called
recursively from mkpath. It is thus calling fs_fsync several times,
before the entire path is constructed.
The fs_fsync function checked that the path existed through open, but
it will still do possibly numerous useless calls depending on the path.

With tee_fs_root being replaced with a tee_fs_fd, we must wait for the
end of tee_supp_fs_init for tee_fs_fd to be valid.

Signed-off-by: Pierre Ducroquet <pinaraf@pinaraf.info>