tee-supplicant: use openat() instead of open() for ree_fs_new_open()
There are several possibilities on a system to temporarily loose access
to the tee data folder when using the absolute path. In my current
setup, this happens when using full disc encryption backed by the fTPM.
Tee-supplicant has to be started early at boot time, with the root
filesystem being mounted later. Thus for short windows of time, the tee
datas are available only when referred to using a file descriptor.
The openat() call (along with fstatat(), renameat() and so on) allows
this, while also being faster and safer against TOCTTOU or similar
attacks/bugs.
This first patch converts only the ree_fs_new_open() function and
introduces the required elements for further conversions.

Signed-off-by: Pierre Ducroquet <pinaraf@pinaraf.info>
Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>