ftrace: Fix not to write duplicate header when append to existing ftrace file
Currently, tee-supplicant appends to an existing prof file if OP-TEE
provides a non-zero id in the rpc call parameter. However, for ftrace
data, there is always a header in the buffer.

This patch makes tee-supplicant not to write duplicate header by
locating the ftrace magic and discarding data before it when
tee-supplicant receives a non-zero ftrace id.

Signed-off-by: Leo Chen <shf.chen@mediatek.com>