tee-supplicant: add missing rule for the mmcblk[0-9]rpmb group
Fix rootless tee-supplicant handling for RPMB use-case. The RPMB
communcation fails badly if the in-kernel RPMB handling was not enabled
and the rootless tee-supplicant is used because the default group for
the /dev/mmcblk[0-9]rpmb is 'root'.

Set the group to same group used for /dev/teepriv[0-9] to fix this and
to allow rootless tee-supplicant usage with legacy user-space RPMB
handling.

Acked-by: Jerome Forissier <jerome.forissier@linaro.org>
Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>