tee-supplicant: add priority prefix for udev rule
The manpage of udev states that "All rule files are read in lexical
order." This coincides with the established convention of assigning a
two-digit priority prefix in front of the respective rule file name.

In case of "optee-udev.rules", we are currently missing that prefix.
When looking for context, it seems suitable to choose a priority in the
realm of the already existing "60-tpm-udev.rules", since that serves a
similar purpose like the rule for optee-client.

This commit hence changes the installed rule file name to
"60-optee-udev.rules", with "60-" being a configurable default value.

Signed-off-by: Holger Assmann <h.assmann@pengutronix.de>
Acked-by: Jerome Forissier <jerome.forissier@linaro.org>